Divide AskPageAIChat into different files
